sl-14: * rewrite protected routes component/fixed errors#1
