+++ frontend/src/index.tsx
@@ -31,15 +31,15 @@ createRoot(document.querySelector('#root') as HTMLElement).render(
                     path: AppRoute.ROOT,
                     element: 'Root',
                   },
-                  {
-                    path: AppRoute.SIGN_IN,
-                    element: <Auth />,
-                  },
-                  {
-                    path: AppRoute.SIGN_UP,
-                    element: <Auth />,
-                  },
                 ],
+              },
+              {
+                path: AppRoute.SIGN_IN,
+                element: <Auth />,
+              },
+              {
+                path: AppRoute.SIGN_UP,
+                element: <Auth />,
               },
             ],
           },
