+++ frontend/src/pages/auth/components/components.tsx
@@ -1,3 +1,3 @@
-export { ProtectRouteComponent } from './protect-route-component/protect-route-component.js';
+export { ProtectRouteComponent } from '../../../libs/components/protect-route-component/protect-route-component.js';
 export { SignInForm } from './sign-in-form/sign-in-form.js';
 export { SignUpForm } from './sign-up-form/sign-up-form.js';
